LayoutTests/svg/animations/animate-end-attribute-expected.txt

SVG 1.1 dynamic animation tests

PASS rect.x.baseVal.value is 100
PASS rect.x.baseVal.value is 50
PASS rect.x.baseVal.value is 200
PASS rect.x.baseVal.value is 200
PASS successfullyParsed is true

TEST COMPLETE
